fix(mobile): change lat and lng type - int to double

+++ apps/mobile/lib/modules/attractions/models/attraction_model.dart
@@ -19,9 +19,9 @@ class Attraction {
 
   String cityId;
 
-  int latitude;
+  double latitude;
 
-  int longitude;
+  double longitude;
 
   String banner;
